requirements.txt
 
python-bitcoinrpc==1.2
